fixing 3.4.2

+++ trunk/algorithm/serialization/src/binary.cpp
@@ -110,7 +110,7 @@ int Serializer::store<std::string>(Base &_rb, FncData &_rfd){
 	idbg("");
 	if(!rs.cpb) return OK;
 	std::string * c = reinterpret_cast<std::string*>(_rfd.p);
-	rs.estk.push(ExtData(c->size()));
+	rs.estk.push(ExtData((uint32)c->size()));
 	rs.replace(FncData(&Serializer::storeBinary, (void*)c->data(), _rfd.n, c->size()));
 	rs.fstk.push(FncData(&Base::popEStack, NULL));
 	rs.fstk.push(FncData(&Serializer::store<uint32>, &rs.estk.top().u32()));
